restructure canopy functionality into putr functions in separate file
